V1
From the post at http://www.mobileread.com/forums/showpost.php?p=1946508&postcount=16

I made an app to enable the no-refresh mode, see norefreshenabler.zip attached
to this post for the apk file and sourcecode. Once installed, just run the app
and the no-refresh mode is enabled automatically. With no-refresh mode i mean
the black and white mode you get when you zoom in the browser. With this app
you do not need to do that trick anymore, just run it.

It is based on the code of Morkl, I edited the EinkListView.java so that it
doesn't need a modified android sdk to compile. It now uses reflection, so the
app searches at runtime for the right method and calls it. For example:
Morkl's code:

	@Override
	public void invalidate() {
		super.invalidate(mUpdateMode);
	}

New:

	@Override
	public void invalidate() {
		try {
			Method invalidateMethod = super.getClass().getMethod("invalidate", int.class);
			invalidateMethod.invoke(this, mUpdateMode);
		} catch(Exception e) {
			e.printStackTrace();
		}
	}

I tried to install/run the app on my phone and that works, it just prints a
stacktrace because the method does not exist.