V2
From the post at http://www.mobileread.com/forums/showpost.php?p=1956700&postcount=33

@MisterNo I made a new version of the norefreshenabler app from post #16. After
you install it, you get 2 apps, one to enable the no-refresh mode and another
to restore the default. There is no close button anymore, it closes itself
automatically. I tried closing the app immediately but when I do that the
screen mode does not change, i added a timer with a 2 second duration and that
works on my reader. Let me know if that time is not enough.

@entodoays As far as I know there is no way to get the current updatemode of
the screen, which is required to create a toggle app.

It might be possible to use the code of the webbrowser and implement it into
the android framework. Unfortunately, I don't have time for that at the moment.
Holding a button to switch modes would be nice indeed. I know you can launch an
app by holding the search button, do you know if you can do this with other
buttons?

I tried setting the mode to 3 in my app, I did that in the Iconizedlistview.zip
in one of my posts earlier in this thread. Updatemode 3 just sets the
updatemode of the listview and not the whole screen so that does not work.